requirements.txt
pywinauto
bs4
requests
demjson
dill
click
six
flask
Pillow
pytesseract
pandas
pyperclip
rqopen-client
easyutils==0.1.7
